requirements.txt
# AI Driven Realtime Log Analyser - Dependencies
# Production-ready package versions

# Core Dependencies
pyyaml>=6.0
python-dateutil>=2.8.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.1.0
joblib>=1.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Real-time Communication
websockets>=11.0.0

# Development and Testing
pytest>=7.1.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0

# Web Framework (for future API)
fastapi>=0.85.0
uvicorn>=0.18.0

# HTTP Requests
requests>=2.28.0

# Utilities
click>=8.1.0
tqdm>=4.64.0
colorama>=0.4.5

# Optional Advanced Features (uncomment if needed)
# tensorflow>=2.10.0
# torch>=1.12.0
# nltk>=3.8
# spacy>=3.4.0
# jira>=3.4.0
